+++ app/models/project.rb
@@ -57,6 +57,7 @@ class Project < ActiveRecord::Base
 
   has_many :domains, -> { where deleted: false }
 
+  has_many :randomizations, -> { where deleted: false }
   has_many :randomization_schemes, -> { where deleted: false }
 
   # Model Methods