+++ test/models/list_option_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class ListOptionTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end